feat(day 5): day 5 in python
This was a little harder than I thought because it turns out that in
part 2, the dependencies ended up being circular. And I was trying to be
clever and create a universal set of flattened dependencies, but if you
do that, then you end up the wrong ordering. This is because in each set
of numbers, not all the numbers are used, therefore, you don't hit the
circular dependencies.

This also gave me a good chance to improve my test harnesses
significantly because of some of the weird cases.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>