feat(day 6): day 6 in python
I'm not 100% happy with this algorithm because it can probably be
optimized. Right now it looks through all of the elements that are
visited and attempts to place an obstacle in those locations and then
runs a complete traverse.

This can probably be optimized by having the traverse and obstacles
nested together, this would not only reduce an extra full traversal, but
for each step it would reduce the need to have to recalculate everything
up to that element.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>