feat(day 6): day 6 in julia
This is, for the most part, a direct port of the algorithm from python,
but I did a little bit of an optimization in the obstacle placement
where I use a matrix, which seems to be much faster for finding the
elements. The end result is that the Julia implementation is about 250%
faster than the Python implementation.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>