feat(debug): add debugging for Julia
This brings in a debugging target for Julia. Because this is a mixed
language workspace with a bunch of different programs in it, I needed
to create my own launch configuration to debug Julia code. Seems like it
works pretty well so far.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>