feat(day 8): day 8 in python
Today was fairly straight forward. The only thing that caught me up on
the first attempt was that I had a messed up boundary condition. It
still is an annoying n^2 algorithm, but whatever - I'm not sure there's
a way to avoid that on something that requires pairwise comparisons.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>