feat(day 9): day 9 in julia
I implemented the defragment differently here - using a list of the free
spaces so it didn't have to crawl everything. The result was a speedup
to 2.45s vs 9.24s. Of note, I left off one part of the code, merging of
newly free space with space after it. But, actually, thinking about the
problem, I didn't need to merge the space before it either, because that
space will be skipped over and nothing will be moved there anyway.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>