feat(day 10): day 10 in python
This was a pretty simple search algorithm across an array. I decided to
make it use numpy because in theory that should be more efficient for a
lot of the use cases. Not sure if it actually made a difference here,
but it was good practice.

The second part of this was actually a degenerate case of the first - in
the sense that the first part required additional code over the second.
In the first part, I needed to keep a visit map to know which cells had
been visited. In the second part, there's no need for a visit map. This
made the second part so easy that I decided just to go for it without
unit tests and it worked -- because it was simple.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>