feat(day 11): day 11 in python
The first part of Day 11 was pretty straight forward. It's trivial to
calculate that many stones in memory. For the second part, I tried to
get clever with various forms of the algorithm and using queues and what
not -- or moving items out of memory once processed, etc. But, then you
realize that the node splits a little more than half the time and
2^(75/2) == big number. That caused me to look at the number of
different values at each step and realize there was only a limited
number of values at each step - so rather than keep track of individual
stones, I just kept track of the numbers in a dict. Easy peasy.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>