feat(day 12): day 12 in Python
This was a pretty easy problem. I decided to do it with a deque (double
ended queue) rather than recursion on the off chance that they had
something sneaky that would blow up the call stack. The hardest part was
counting the sides and figuring out the edge cases (heh) for when to
count a new side.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>