feat(day 14): day 14 in python
The trick here was in part 2 to figure out how to make an algorithm
detect a vaguely specified picture. Reasoning that the picture probably
had lower variance, I made it so it displayed the map only if the
variance of that image was less the least variance previously seen. This
seems to work well and it's also what some other folks on
/r/adventofcode came up with.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>