feat(day 16): day 16 in python
The first part of the problem was a pretty simple graph traversal. The
second part of the problem was a little more challenging because you
needed to record the score for each direction a path could come into the
cell. This probably resulted in _MANY_ loops. I could probably fix this
by doing a check to make sure they're within 1000 or something like
that.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>