feat(day 17): day 17 in python
The first part was pretty easy and straightforward. The second part
required a bit of program analysis to understand. The key was seeing
that everyhing was controlled by powers of 8. The length of the input
value had to be somewhere between 8**15 and 8**16 (because the input was
16 characters long). From there it was just iterating over the each
character to find the right one while decrementing the power of eight.
This is way faster than brute forcing it.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>